LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/intercepted-referrer.https-expected.txt

Service Worker intercepted navigation preserves document.referrer


PASS document.referrer for a navigation intercepted by a Service Worker is unchanged and matches the parent URL
